requirements.txt
Django
psycopg2-binary
django-cors-headers
djangorestframework
Pillow